refactor: Update dependencies in docs cache, navigation links, and PaymentElementDemo component import
