+++ docs/.vitepress/cache/deps/vue.js
@@ -158,8 +158,8 @@ import {
   withMemo,
   withModifiers,
   withScopeId
-} from "./chunk-ATHFSSNY.js";
-import "./chunk-CSAU5B4Q.js";
+} from "./chunk-OXQEVQFY.js";
+import "./chunk-PZ5AY32C.js";
 export {
   BaseTransition,
   BaseTransitionPropsValidators,
