+++ docs/.vitepress/theme/index.js
@@ -1,3 +1,5 @@
+import './tailwind.postcss'
+
 // .vitepress/theme/index.js
 import DefaultTheme from 'vitepress/theme'
 import './custom.css'
