+++ docs/.vitepress/theme/tailwind.postcss
@@ -1,0 +1,5 @@
+@tailwind base;
+
+@tailwind components;
+
+@tailwind utilities;
