--- docs/components/advanced-ui-element.md
@@ -1,1 +1,0 @@
-# The Advanced UI Element Component
