--- docs/composables/use-element.md
@@ -1,1 +1,0 @@
-# The useElement Composable
