--- docs/composables/use-elements.md
@@ -1,1 +1,0 @@
-# The useElements Composable
