+++ docs/faqs/client-secret.md
@@ -1,0 +1,9 @@
+# What is a client secret?
+
+A client secret is a unique key that is used to authenticate a payment intent. It is generated by the Stripe API and is used to confirm that the payment intent is valid.
+
+Take note that this is different from the publishable key, which is used to authenticate the client-side of the payment process.
+
+To get the client secret, you can create a payment intent using the Stripe API and retrieve the client secret from the response.
+
+Here's a complete documentation about payment intent and client secret: [Payment Intents](https://docs.stripe.com/api/payment_intents/object#payment_intent_object-client_secret)
